feat : 네비바 카테고리 이동 기능 추가
